+++ server/standalone/dbal/datasample.go
@@ -3,6 +3,7 @@ package dbal
 import (
 	"errors"
 	"strconv"
+	"time"
 
 	sq "github.com/Masterminds/squirrel"
 	"github.com/go-playground/log/v7"
@@ -11,58 +12,156 @@ import (
 	"github.com/owkin/orchestrator/lib/asset"
 	"github.com/owkin/orchestrator/lib/common"
 	orcerrors "github.com/owkin/orchestrator/lib/errors"
-	"google.golang.org/protobuf/encoding/protojson"
+	"google.golang.org/protobuf/types/known/timestamppb"
 )
+
+type sqlDataSample struct {
+	Key             string
+	Owner           string
+	TestOnly        bool
+	Checksum        string
+	CreationDate    time.Time
+	DataManagerKeys []string
+}
+
+func (ds *sqlDataSample) toDataSample() *asset.DataSample {
+	return &asset.DataSample{
+		Key:             ds.Key,
+		DataManagerKeys: ds.DataManagerKeys,
+		Owner:           ds.Owner,
+		TestOnly:        ds.TestOnly,
+		Checksum:        ds.Checksum,
+		CreationDate:    timestamppb.New(ds.CreationDate),
+	}
+}
 
 // DataSampleExists implements persistence.DataSampleDBAL
 func (d *DBAL) DataSampleExists(key string) (bool, error) {
-	row := d.tx.QueryRow(d.ctx, `select count(id) from "datasamples" where id=$1 and channel=$2`, key, d.channel)
+	stmt := getStatementBuilder().
+		Select("COUNT(key)").
+		From("datasamples").
+		Where(sq.Eq{"channel": d.channel, "key": key})
+
+	row, err := d.queryRow(stmt)
+	if err != nil {
+		return false, err
+	}
 
 	var count int
-	err := row.Scan(&count)
+	err = row.Scan(&count)
 
 	return count == 1, err
 }
 
-// AddDataSamples insert samples in storage in batch mode.
+// AddDataSamples add one or multiple data samples to storage.
 func (d *DBAL) AddDataSamples(datasamples ...*asset.DataSample) error {
 	log.WithField("numSamples", len(datasamples)).Debug("dbal: adding multiple datasamples in batch mode")
+	err := d.insertDataSamples(datasamples)
+	if err != nil {
+		return err
+	}
 
+	return d.insertDataSampleDataManagers(datasamples...)
+}
+
+// insertDataSamples insert data samples in database in batch mode.
+func (d *DBAL) insertDataSamples(datasamples []*asset.DataSample) error {
 	_, err := d.tx.CopyFrom(
 		d.ctx,
 		pgx.Identifier{"datasamples"},
-		[]string{"id", "asset", "channel"},
+		[]string{"key", "channel", "owner", "test_only", "checksum", "creation_date"},
 		pgx.CopyFromSlice(len(datasamples), func(i int) ([]interface{}, error) {
-			v, err := protojson.Marshal(datasamples[i])
-			if err != nil {
-				return nil, err
-			}
+			ds := datasamples[i]
+
 			// expect binary representation, not string
-			id, err := uuid.Parse(datasamples[i].Key)
+			key, err := uuid.Parse(ds.Key)
 			if err != nil {
 				return nil, err
 			}
-			return []interface{}{id, v, d.channel}, nil
+
+			return []interface{}{key, d.channel, ds.Owner, ds.TestOnly, ds.Checksum, ds.CreationDate.AsTime()}, nil
 		}),
 	)
 
 	return err
 }
 
+// insertDataSampleDataManagers insert the datasample-datamanager relations in database in batch mode.
+func (d *DBAL) insertDataSampleDataManagers(datasamples ...*asset.DataSample) error {
+	rows := make([][]interface{}, 0)
+
+	for _, ds := range datasamples {
+		if ds.DataManagerKeys != nil {
+			dataSampleKey, err := uuid.Parse(ds.Key)
+			if err != nil {
+				return err
+			}
+
+			for _, dmKey := range ds.DataManagerKeys {
+				dataManagerKey, err := uuid.Parse(dmKey)
+				if err != nil {
+					return err
+				}
+				rows = append(rows, []interface{}{dataSampleKey, dataManagerKey})
+			}
+		}
+	}
+
+	_, err := d.tx.CopyFrom(
+		d.ctx,
+		pgx.Identifier{"datasample_datamanagers"},
+		[]string{"datasample_key", "datamanager_key"},
+		pgx.CopyFromRows(rows),
+	)
+
+	return err
+}
+
+func (d *DBAL) deleteDataSampleDataManagers(dataSampleKey string) error {
+	stmt := getStatementBuilder().
+		Delete("datasample_datamanagers").
+		Where(sq.Eq{"datasample_key": dataSampleKey})
+
+	return d.exec(stmt)
+}
+
 // UpdateDataSample implements persistence.DataSampleDBAL
 func (d *DBAL) UpdateDataSample(dataSample *asset.DataSample) error {
-	stmt := `update "datasamples" set asset=$3 where id=$1 and channel=$2`
-	_, err := d.tx.Exec(d.ctx, stmt, dataSample.GetKey(), d.channel, dataSample)
-	return err
+	err := d.deleteDataSampleDataManagers(dataSample.Key)
+	if err != nil {
+		return err
+	}
+
+	err = d.insertDataSampleDataManagers(dataSample)
+	if err != nil {
+		return err
+	}
+
+	stmt := getStatementBuilder().
+		Update("datasamples").
+		Set("owner", dataSample.Owner).
+		Set("test_only", dataSample.TestOnly).
+		Set("checksum", dataSample.Checksum).
+		Where(sq.Eq{"channel": d.channel, "key": dataSample.Key})
+
+	return d.exec(stmt)
 }
 
 // GetDataSample implements persistence.DataSample
 func (d *DBAL) GetDataSample(key string) (*asset.DataSample, error) {
-	row := d.tx.QueryRow(d.ctx, `select "asset" from "datasamples" where id=$1 and channel=$2`, key, d.channel)
+	stmt := getStatementBuilder().
+		Select("key", "owner", "test_only", "checksum", "creation_date", "datamanager_keys").
+		From("expanded_datasamples").
+		Where(sq.Eq{"channel": d.channel, "key": key})
 
-	datasample := new(asset.DataSample)
-	err := row.Scan(datasample)
+	row, err := d.queryRow(stmt)
+	if err != nil {
+		return nil, err
+	}
 
+	ds := new(sqlDataSample)
+
+	err = row.Scan(&ds.Key, &ds.Owner, &ds.TestOnly, &ds.Checksum, &ds.CreationDate, &ds.DataManagerKeys)
 	if err != nil {
 		if errors.Is(err, pgx.ErrNoRows) {
 			return nil, orcerrors.NewNotFound("datasample", key)
@@ -70,7 +169,7 @@ func (d *DBAL) GetDataSample(key string) (*asset.DataSample, error) {
 		return nil, err
 	}
 
-	return datasample, nil
+	return ds.toDataSample(), nil
 }
 
 // QueryDataSamples implements persistence.DataSample
@@ -81,16 +180,16 @@ func (d *DBAL) QueryDataSamples(p *common.Pagination, filter *asset.DataSampleQu
 	}
 
 	stmt := getStatementBuilder().
-		Select("asset").
-		From("datasamples").
+		Select("key", "owner", "test_only", "checksum", "creation_date", "datamanager_keys").
+		From("expanded_datasamples").
 		Where(sq.Eq{"channel": d.channel}).
-		OrderByClause("asset->>'creationDate' asc, id").
+		OrderByClause("creation_date ASC, key").
 		Offset(uint64(offset)).
 		// Fetch page size + 1 elements to determine whether there is a next page
 		Limit(uint64(p.Size + 1))
 
 	if filter != nil && len(filter.Keys) > 0 {
-		stmt = stmt.Where(sq.Eq{"id": filter.Keys})
+		stmt = stmt.Where(sq.Eq{"key": filter.Keys})
 	}
 
 	rows, err := d.query(stmt)
@@ -103,21 +202,21 @@ func (d *DBAL) QueryDataSamples(p *common.Pagination, filter *asset.DataSampleQu
 	var count int
 
 	for rows.Next() {
-		datasample := new(asset.DataSample)
+		ds := new(sqlDataSample)
 
-		err = rows.Scan(datasample)
+		err = rows.Scan(&ds.Key, &ds.Owner, &ds.TestOnly, &ds.Checksum, &ds.CreationDate, &ds.DataManagerKeys)
 		if err != nil {
 			return nil, "", err
 		}
 
-		datasamples = append(datasamples, datasample)
+		datasamples = append(datasamples, ds.toDataSample())
 		count++
 
 		if count == int(p.Size) {
 			break
 		}
 	}
-	if err := rows.Err(); err != nil {
+	if err = rows.Err(); err != nil {
 		return nil, "", err
 	}
 
@@ -130,19 +229,17 @@ func (d *DBAL) QueryDataSamples(p *common.Pagination, filter *asset.DataSampleQu
 	return datasamples, bookmark, nil
 }
 
-// GetDataSampleKeysByManager returns sample keys linked to given manager.
+// GetDataSampleKeysByManager returns sample keys linked to a given manager.
 func (d *DBAL) GetDataSampleKeysByManager(dataManagerKey string, testOnly bool) ([]string, error) {
-	testOnlyFilter := `not`
-	if testOnly {
-		testOnlyFilter = ``
-	}
+	stmt := getStatementBuilder().
+		Select("datasample_key").
+		From("datasample_datamanagers").
+		Join("datasamples ds ON ds.key = datasample_datamanagers.datasample_key").
+		Where(sq.Eq{"ds.test_only": testOnly}).
+		Where(sq.Eq{"datamanager_key": dataManagerKey}).
+		OrderByClause("creation_date ASC, key")
 
-	query := `select "id" from "datasamples" where channel=$1 and (asset->'dataManagerKeys') ? $2 and ` +
-		testOnlyFilter +
-		` (asset ? 'testOnly' and (asset->'testOnly')::boolean) order by asset->>'creationDate' asc, id`
-
-	rows, err := d.tx.Query(d.ctx, query, d.channel, dataManagerKey)
-
+	rows, err := d.query(stmt)
 	if err != nil {
 		return nil, err
 	}
@@ -160,7 +257,7 @@ func (d *DBAL) GetDataSampleKeysByManager(dataManagerKey string, testOnly bool) 
 		}
 		datasampleKeys = append(datasampleKeys, datasampleKey)
 	}
-	if err := rows.Err(); err != nil {
+	if err = rows.Err(); err != nil {
 		return nil, err
 	}