+++ server/standalone/dbal/datasample_test.go
@@ -18,7 +18,8 @@ func TestGetDataSampleFail(t *testing.T) {
 	mock.ExpectBegin()
 
 	uid := "4c67ad88-309a-48b4-8bc4-c2e2c1a87a83"
-	mock.ExpectQuery(`select "asset" from "datasamples" where id=`).WithArgs(uid, testChannel)
+	mock.ExpectQuery(`SELECT key, owner, test_only, checksum, creation_date, datamanager_keys FROM expanded_datasamples`).
+		WithArgs(testChannel, uid)
 
 	tx, err := mock.Begin(context.Background())
 	require.NoError(t, err)